Reset loadgen meters on clearmetrics (#3630) (#3635)
* Reset loadgen meters on clearmetrics

clearmetrics was a no-op, so Supercluster's max-TPS binary search (which calls
clearmetrics before every step) saw loadgen_txn_attempted accumulate across
steps and a stale loadgen_run_failed linger after a failed step. Reset the
loadgen domain counters via absolute(0), matching stellar-core's resettable
meters.

Fixes #3630


* Regression test for #3630 — loadgen clearmetrics reset is a no-op

Adds reset_loadgen_meters() (currently using CounterMetric::absolute(0),
which is AtomicU64::fetch_max(0) — a no-op on any non-zero counter) and a
render-level regression test that increments the loadgen meters then calls
the reset and asserts every loadgen series renders 0. Fails on the current
fetch_max mechanism (renders the accumulated value), proving #3630.

Refs #3630


* Make loadgen clearmetrics reset actually render 0

The prior fix used CounterMetric::absolute(0), which bottoms out at
AtomicU64::fetch_max(0) in the metrics crate — a no-op on any non-zero
counter, so the loadgen meters were never reset and #3630 stayed broken
(the reviewer proved this with a render-level probe).

A monotonic Prometheus counter cannot be lowered through any public API of
the installed metrics-exporter-prometheus recorder (the backing
Arc<AtomicU64> is unreachable, and absolute = fetch_max). The only primitive
that yields a true rendered 0 is a gauge's set (a true swap), which also
matches stellar-core's resettable medida meters (MetricResetter,
ApplicationImpl.cpp:1375).

So model the 15 loadgen meters as pre-registered gauges and reset them with
set(0.0) in reset_loadgen_meters(), invoked from the clearmetrics HTTP path
only. Supercluster reads the numeric value via the compat /metrics JSON
(LOADGEN_COMPAT_MAP -> count), which is type-agnostic, so the counter->gauge
exposition change is invisible to the consumer. Increment call sites
(loadgen.rs, main.rs) switch to f64; the dotted compat-key mapping and the
domain gate (empty || "loadgen") are unchanged.

The regression test test_clear_loadgen_meters_resets_to_zero_rendered now
passes: it renders N, calls reset_loadgen_meters(), and asserts every
loadgen series renders exactly 0.

Refs #3630


---------

+++ crates/simulation/src/loadgen.rs
@@ -1637,7 +1637,7 @@ impl LoadGenerator {
 
             // One step per loop iteration (parity: stellar-core
             // LoadGenerator::getTxPerStep() mStepMeter.Mark(), LoadGenerator.cpp:211).
-            app_metrics::LOADGEN_STEP_COUNT.increment(1);
+            app_metrics::LOADGEN_STEP_COUNT.increment(1.0);
 
             // Check if all transactions for the current phase are submitted
             if !config.are_txs_remaining() {
@@ -2080,13 +2080,13 @@ impl LoadGenerator {
             // (stellar-core LoadGenerator.cpp:874).
             if config.mode == LoadGenMode::PayPregenerated {
                 if !matches!(result, TxQueueResult::Added) {
-                    app_metrics::LOADGEN_TXN_REJECTED.increment(1);
+                    app_metrics::LOADGEN_TXN_REJECTED.increment(1.0);
                 }
                 return matches!(result, TxQueueResult::Added);
             }
 
             if !matches!(result, TxQueueResult::Added) {
-                app_metrics::LOADGEN_TXN_REJECTED.increment(1);
+                app_metrics::LOADGEN_TXN_REJECTED.increment(1.0);
             }
 
             // `txBAD_SEQ` and `QueueFull`/`TryAgainLater` retries each consume
@@ -2751,41 +2751,42 @@ fn mark_tx_meters(mode: LoadGenMode, env: &TransactionEnvelope) {
     let xdr_size = envelope_xdr_size(env);
     match mode {
         LoadGenMode::Pay | LoadGenMode::PayPregenerated => {
-            app_metrics::LOADGEN_PAYMENT_SUBMITTED.increment(envelope_num_operations(env));
-            app_metrics::LOADGEN_PAYMENT_BYTES.increment(xdr_size);
+            app_metrics::LOADGEN_PAYMENT_SUBMITTED.increment(envelope_num_operations(env) as f64);
+            app_metrics::LOADGEN_PAYMENT_BYTES.increment(xdr_size as f64);
         }
         LoadGenMode::SorobanUpload => {
-            app_metrics::LOADGEN_SOROBAN_UPLOAD.increment(1);
+            app_metrics::LOADGEN_SOROBAN_UPLOAD.increment(1.0);
         }
         LoadGenMode::SorobanInvokeSetup => {
-            app_metrics::LOADGEN_SOROBAN_SETUP_INVOKE.increment(1);
+            app_metrics::LOADGEN_SOROBAN_SETUP_INVOKE.increment(1.0);
         }
         LoadGenMode::SorobanUpgradeSetup => {
-            app_metrics::LOADGEN_SOROBAN_SETUP_UPGRADE.increment(1);
+            app_metrics::LOADGEN_SOROBAN_SETUP_UPGRADE.increment(1.0);
         }
         LoadGenMode::SorobanInvoke | LoadGenMode::SorobanInvokeApplyLoad => {
-            app_metrics::LOADGEN_SOROBAN_INVOKE.increment(1);
+            app_metrics::LOADGEN_SOROBAN_INVOKE.increment(1.0);
         }
         LoadGenMode::SorobanCreateUpgrade => {
-            app_metrics::LOADGEN_SOROBAN_CREATE_UPGRADE.increment(1);
+            app_metrics::LOADGEN_SOROBAN_CREATE_UPGRADE.increment(1.0);
         }
         LoadGenMode::MixedClassicSoroban => {
             // Sub-mode is decided per-tx; classify by inspecting the built tx
             // (parity: core's execute() switch on mLastMixedMode, equivalently
             // the MIXED_PREGEN_* isSoroban() branch).
             if envelope_is_soroban(env) {
-                app_metrics::LOADGEN_SOROBAN_INVOKE.increment(1);
+                app_metrics::LOADGEN_SOROBAN_INVOKE.increment(1.0);
             } else {
-                app_metrics::LOADGEN_PAYMENT_SUBMITTED.increment(envelope_num_operations(env));
-                app_metrics::LOADGEN_PAYMENT_BYTES.increment(xdr_size);
+                app_metrics::LOADGEN_PAYMENT_SUBMITTED
+                    .increment(envelope_num_operations(env) as f64);
+                app_metrics::LOADGEN_PAYMENT_BYTES.increment(xdr_size as f64);
             }
         }
     }
 
     // Per-tx attempt + bytes (parity: txm.mTxnAttempted.Mark() +
     // txm.mTxnBytes.Mark(...) at the end of execute()).
-    app_metrics::LOADGEN_TXN_ATTEMPTED.increment(1);
-    app_metrics::LOADGEN_TXN_BYTES.increment(xdr_size);
+    app_metrics::LOADGEN_TXN_ATTEMPTED.increment(1.0);
+    app_metrics::LOADGEN_TXN_BYTES.increment(xdr_size as f64);
 }
 
 fn envelope_seq_num(env: &TransactionEnvelope) -> i64 {
